+++ samples/BaGetWebApplication/appsettings.json
@@ -3,18 +3,6 @@
     "ConnectionString": "Data Source=baget.db"
   },
 
-  "Storage": {
-    "TODO":  "Make storage field optional"
-  },
-
-  "Search": {
-    "TODO":  "Make search field optional"
-  },
-
-  "Mirror": {
-    "TODO":  "Make mirror field optional."
-  },
-
   "Logging": {
     "LogLevel": {
       "Default": "Information",
